train-company-ticketing-report/requirements.txt
psycopg[binary]==3.2.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
